Move CA1018 to its new location
