--- PortFxCop/src/Microsoft.ApiDesignGuidelines.Analyzers/VisualBasic/BasicMarkAttributesWithAttributeUsage.Fixer.vb
@@ -1,19 +1,0 @@
-' Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
-
-Imports System.Composition
-Imports System.Linq
-Imports Microsoft.CodeAnalysis
-Imports Microsoft.CodeAnalysis.CodeFixes
-Imports Microsoft.CodeAnalysis.VisualBasic
-Imports Microsoft.CodeAnalysis.VisualBasic.Syntax
-
-Namespace Microsoft.ApiDesignGuidelines.Analyzers     
-    ''' <summary>
-    ''' CA1018: Mark attributes with AttributeUsageAttribute
-    ''' </summary>
-    <ExportCodeFixProvider(LanguageNames.VisualBasic), [Shared]>
-    Public NotInheritable Class BasicMarkAttributesWithAttributeUsageFixer
-        Inherits MarkAttributesWithAttributeUsageFixer 
-
-    End Class
-End Namespace
